+++ taxi_environment.py
@@ -3,10 +3,13 @@
 import gym
 from gym.utils import seeding
 import numpy as np
-from multitaxienv.config import TAXI_ENVIRONMENT_REWARDS, BASE_AVAILABLE_ACTIONS, ALL_ACTIONS_NAMES
+from config import TAXI_ENVIRONMENT_REWARDS, BASE_AVAILABLE_ACTIONS, ALL_ACTIONS_NAMES
 from gym.spaces import Box, Tuple, MultiDiscrete
-#from multitaxienv.utils import rendering_utils, basic_utils, actions_utils, \
-#    observation_utils, reward_utils
+import random
+import sys
+from contextlib import closing
+from io import StringIO
+from gym import utils
 
 orig_MAP = [
     "+---------+",
@@ -50,45 +53,28 @@ simple_MAP = [
 ]
 
 
-# TODO change documentation to refer to rllib
-# TODO - Create requirements.txt
-# TODO Create another notebook for all the changes in the environment
-
-# supporting rllib and gym via this alternating variable
-environment_type = 'gym'
-if environment_type == 'rllib':
-    from ray.rllib.env import MultiAgentEnv
-    environment_to_derive_from = MultiAgentEnv
-else:
-    import gym
-    environment_to_derive_from = gym.Env
-
-
-class TaxiEnv(environment_to_derive_from):
+class TaxiEnv:
     """
     The Taxi Problem
     from "Hierarchical Reinforcement Learning with the MAXQ Value Function Decomposition"
-    by Tom Dietterich, adjusted for MARL by Ofir Abu
+    by Tom Dietterich
     Description:
-    There are designated sources and destinations of passengers (chosen randomly).
-    When the episode starts, the taxis start off at a random square and the passenger is at a random location.
-    The taxis drive to the passenger's location, picks up the passenger, drives to the passenger's destination
+    There are four designated locations in the grid world indicated by R(ed), G(reen), Y(ellow), and B(lue).
+    When the episode starts, the taxi starts off at a random square and the passenger is at a random location.
+    The taxi drives to the passenger's location, picks up the passenger, drives to the passenger's destination
     (another one of the four specified locations), and then drops off the passenger. Once the passenger is dropped off,
     the episode ends.
-    A taxi can pickup more than one passenger, capacity is an hyperparameter.
 
     Observations:
     A list (taxis, fuels, pass_start, destinations, pass_locs):
         taxis:                  a list of coordinates of each taxi
         fuels:                  a list of fuels for each taxi
-        pass_start:             a list of starting coordinates for each passenger (current position or last available)
+        pass_start:             a list of starting coordinates for taeach passenger (current position or last available)
         destinations:           a list of destination coordinates for each passenger
         passengers_locations:   a list of locations of each passenger.
                                 -1 means delivered
                                 0 means not picked up
                                 positive number means the passenger is in the corresponding taxi number
-
-    Observations consists of the pixel map around the taxi and a status vector.
 
     Passenger start: coordinates of each of these
     - -1: In a taxi
@@ -109,7 +95,7 @@ class TaxiEnv(environment_to_derive_from):
     - 3: B(lue)
 
     Fuel:
-     - 0 to np.inf: default with 10, 0 means infinity fuel amount
+     - 0 to np.inf: default with 10
 
     Actions:
     Actions are given as a list, each element referring to one taxi's action. Each taxi has 7 actions:
@@ -140,9 +126,10 @@ class TaxiEnv(environment_to_derive_from):
     metadata = {'render.modes': ['human', 'ansi']}
 
     def __init__(self, _=0, num_taxis: int = 1, num_passengers: int = 1, max_fuel: list = None,
-                 domain_map: list = simple_MAP, taxis_capacity: list = None, collision_sensitive_domain: bool = True,
-                 fuel_type_list: list = None, option_to_stand_by: bool = False, view_len: int = 1):
+                 domain_map: list = None, taxis_capacity: list = None, collision_sensitive_domain: bool = True,
+                 fuel_type_list: list = None, option_to_stand_by: bool = False):
         """
+        TODO -  later version make number of passengers dynamic, even in runtime
         Args:
             num_taxis: number of taxis in the domain
             num_passengers: number of passengers occupying the domain at initiailaization
@@ -152,24 +139,21 @@ class TaxiEnv(environment_to_derive_from):
             collision_sensitive_domain: is the domain show and react (true) to collisions or not (false)
             fuel_type_list: list of fuel types of each taxi
             option_to_stand_by: can taxis simply stand in place
-            view_len: width and height of the observation window for a taxi
         """
-
         # Initializing default values
-        self.is_infinite_fuel = False
+        self.num_taxis = num_taxis
         if max_fuel is None:
-            self.max_fuel = [0] * num_taxis
-            self.is_infinite_fuel = True
+            self.max_fuel = [100] * num_taxis  # TODO - needs to figure out how to insert np.inf into discrete obs.space
         else:
             self.max_fuel = max_fuel
 
         if domain_map is None:
-            self.desc = np.asarray(MAP, dtype='c')
+            self.desc = np.asarray(orig_MAP, dtype='c')
         else:
             self.desc = np.asarray(domain_map, dtype='c')
 
         if taxis_capacity is None:
-            self.taxis_capacity = [2] * num_passengers
+            self.taxis_capacity = [1] * num_passengers
         else:
             self.taxis_capacity = taxis_capacity
 
@@ -188,44 +172,46 @@ class TaxiEnv(environment_to_derive_from):
         self.fuel_station2 = None
         self.fuel_stations = []
 
-        self.fuel_stations = rendering_utils.parse_fuel_stations_from_map(self.desc)
+        # initializing map with passengers and fuel stations
+        for i, row in enumerate(self.desc[1:-1]):
+            for j, char in enumerate(row[1:-1:2]):
+                loc = [i, j]
+                if char == b'X':
+                    self.passengers_locations.append(loc)
+                elif char == b'F':
+                    self.fuel_station1 = loc
+                    self.fuel_stations.append(loc)
+                elif char == b'G':
+                    self.fuel_station2 = loc
+                    self.fuel_stations.append(loc)
 
-        self.num_passengers = num_passengers
+        self.coordinates = [[i, j] for i in range(num_rows) for j in range(num_columns)]
 
-        self.coordinates = np.array([[i, j] for i in range(num_rows) for j in range(num_columns)])
-
-        self.passengers_locations = list(self.coordinates[
-            np.random.choice(self.coordinates.shape[0], size=2*num_passengers, replace=False), :])
-
-        self.coordinates = basic_utils.get_array_as_list_type(self.coordinates)
-        self.passengers_locations = basic_utils.get_array_as_list_type(self.passengers_locations)
-
-        self.num_taxis = num_taxis
-        self.taxis_names = ["taxi_" + str(index + 1) for index in range(num_taxis)]
+        # self.num_taxis = num_taxis
+        self.taxis_names = ["taxi_" + str(index + 1) for index in range(self.num_taxis)]
 
         self.collision_sensitive_domain = collision_sensitive_domain
 
         # Indicator list of 1's (collided) and 0's (not-collided) of all taxis
-        self.collided = np.zeros(num_taxis)
+        self.collided = np.zeros(self.num_taxis)
 
         self.option_to_standby = option_to_stand_by
 
         # A list to indicate whether the engine of taxi i is on (1) or off (0), all taxis start as on.
-        self.engine_status_list = list(np.ones(num_taxis).astype(bool))
+        self.engine_status_list = list(np.ones(self.num_taxis).astype(bool))
+
+        self.num_passengers = num_passengers
 
         # Available actions in relation to all actions based on environment parameters.
         self.available_actions_indexes, self.index_action_dictionary, self.action_index_dictionary \
             = self._set_available_actions_dictionary()
         self.num_actions = len(self.available_actions_indexes)
         self.action_space = gym.spaces.Discrete(self.num_actions)
-        # self.obs_space = gym.spaces.MultiDiscrete(self._get_observation_space_list())
-        self.view_len = view_len
-        self.observation_space = Tuple([Box(low=0.0, high=255.0, shape=(2 * self.view_len + 1,
-                                                                        2 * self.view_len + 1, 3), dtype=np.float32),
-                                        MultiDiscrete(self._get_observation_space_list())])
+        self.observation_space = gym.spaces.MultiDiscrete(self._get_observation_space_list())
         self.bounded = False
 
         self.last_action = None
+        self.num_states = self._get_num_states()
 
         self._seed()
         self.state = None
@@ -235,7 +221,17 @@ class TaxiEnv(environment_to_derive_from):
         self.np_random = None
         self.reset()
 
-    # TODO change this function when working on the status observation vector
+    def _get_num_states(self):
+        map_dim = (self.num_rows * self.num_columns)
+        passengers_loc_dim = 1
+        for i in range(self.num_passengers):
+            passengers_loc_dim *= len(self.passengers_locations) + self.num_taxis - i
+        passengers_dest_dim = 1
+        for i in range(self.num_passengers):
+            passengers_dest_dim *= len(self.passengers_locations) - i
+        num_states = map_dim * passengers_loc_dim * passengers_dest_dim
+        return num_states
+
     def _get_observation_space_list(self) -> list:
         """
         Returns a list that emebed the observation space size in each dimension.
@@ -260,10 +256,10 @@ class TaxiEnv(environment_to_derive_from):
         passengers_status_size = [self.num_taxis + 3]
         dimensions_sizes = []
 
-        # for _ in range(self.num_taxis):
-        dimensions_sizes += locations_sizes
-        # for _ in range(self.num_taxis):
-        dimensions_sizes += fuel_size
+        for _ in range(self.num_taxis):
+            dimensions_sizes += locations_sizes
+        for _ in range(self.num_taxis):
+            dimensions_sizes += fuel_size
 
         for _ in range(self.num_passengers):
             dimensions_sizes += 2 * locations_sizes
@@ -272,11 +268,10 @@ class TaxiEnv(environment_to_derive_from):
 
         return [dimensions_sizes]
 
-    # When running trials and finishing exploration - set seed to None
-    def _seed(self, seed=42) -> np.ndarray:
+    def _seed(self, seed=None) -> list:
         """
         Setting a seed for the random sample state generation.
-        Args:list
+        Args:
             seed: seed to use
 
         Returns: list[seed]
@@ -296,37 +291,24 @@ class TaxiEnv(environment_to_derive_from):
             - all engines turn on.
         Args:
 
-        Returns: The reset state (dictionary as in RLlib format).
+        Returns: The reset state.
 
         """
         # reset taxis locations
-        taxis_locations = np.array(self.coordinates)[np.random.choice(range(len(self.coordinates)), self.num_taxis, replace=False), :]
-        taxis_locations = basic_utils.get_array_as_list_type(taxis_locations)
+        taxis_locations = random.sample(self.coordinates, self.num_taxis)
         self.collided = np.zeros(self.num_taxis)
         self.bounded = False
-        self.window_size = self.view_len
+        self.window_size = 5
         self.counter = 0
 
         # refuel everybody
         fuels = [self.max_fuel[i] for i in range(self.num_taxis)]
 
         # reset passengers
-        self.coordinates = np.array(self.coordinates)
-        self.passengers_locations = self.coordinates[
-                                         np.random.choice(self.coordinates.shape[0], size=2 * self.num_passengers,
-                                                          replace=False), :]
-
-        chosen_indices = np.random.choice(self.passengers_locations.shape[0], size=self.num_passengers,
-                                                          replace=False)
-
-        passengers_start_location = self.passengers_locations[chosen_indices, :]
-        passengers_destinations = self.passengers_locations[
-            list(set(range(self.passengers_locations.shape[0])) - set(list(chosen_indices))), :]
-
-        self.coordinates = basic_utils.get_array_as_list_type(self.coordinates)
-        passengers_start_location = basic_utils.get_array_as_list_type(passengers_start_location)
-        passengers_destinations = basic_utils.get_array_as_list_type(passengers_destinations)
-        self.passengers_locations = basic_utils.get_array_as_list_type(self.passengers_locations)
+        passengers_start_location = [start for start in
+                                     random.choices(self.passengers_locations, k=self.num_passengers)]
+        passengers_destinations = [random.choice([x for x in self.passengers_locations if x != start])
+                                   for start in passengers_start_location]
 
         # Status of each passenger: delivered (1), in_taxi (positive number>2), waiting (2)
         passengers_status = [2 for _ in range(self.num_passengers)]
@@ -339,55 +321,16 @@ class TaxiEnv(environment_to_derive_from):
         # resetting dones
         self.dones = {taxi_id: False for taxi_id in self.taxis_names}
         self.dones['__all__'] = False
+        obs = {}
+        for taxi_id in self.taxis_names:
+            obs[taxi_id] = self.get_observation(self.state, taxi_id)
 
-        # setting observation according to the reset state and the rllib dictionary format
-        observations = {}
-        for i, taxi_id in enumerate(self.taxis_names):
-            # obs[taxi_id] = self.get_observation(self.state, taxi_id)
-            observations[taxi_id] = (observation_utils.get_image_obs_by_agent_id(i, self.state, self.num_taxis,
-                                                                                 self.collided, self.view_len,
-                                                                                 self.desc.copy()),
-                                     observation_utils.get_status_vector_observation(self.state,
-                                                                                     taxi_id,
-                                                                                     self.taxis_names,
-                                                                                     self.num_taxis))
-
-        return observations
-
-    def _add_custom_dropoff_for_passengers(self) -> (list, list):
-        """
-        Change the basic action list from config to have dropoff action for each specific passenger.
-        for example, if we have 2 passengers, we'll have 2 seperate actions: "dropoff0", "dropoff1".
-        Returns: adjusted ALL_ACTIONS_NAMES, BASE_AVAILABLE_ACTIONS lists
-
-        """
-
-        def add_custom_dropoff(list_of_actions: list) -> list:
-            """
-            A sub method to add the specific new actions (dropoffs) to a given action list.
-            Args:
-                list_of_actions: the action list to add on the new actions
-
-            Returns: modified action list
-
-            """
-            temp_list = []
-            for action in list_of_actions:
-                if action != "dropoff":
-                    temp_list.append(action)
-                else:
-                    for passenger_id in range(self.num_passengers):
-                        temp_list.append(action + str(passenger_id))
-
-            return temp_list
-
-        new_base_action_list, new_all_action_list = add_custom_dropoff(BASE_AVAILABLE_ACTIONS), \
-                                                    add_custom_dropoff(ALL_ACTIONS_NAMES)
-
-        return new_all_action_list, new_base_action_list
+        return obs
 
     def _set_available_actions_dictionary(self) -> (list, dict, dict):
         """
+
+        TODO: Later versions - maybe return an action-dictionary for each taxi individually.
 
         Generates list of all available actions in the parametrized domain, index->action dictionary to decode.
         Generation is based on the hyper-parameters passed to __init__ + parameters defined in config.py
@@ -396,7 +339,7 @@ class TaxiEnv(environment_to_derive_from):
         (action -> index).
 
         """
-        ALL_ACTIONS_NAMES, BASE_AVAILABLE_ACTIONS = self._add_custom_dropoff_for_passengers()
+
         action_names = ALL_ACTIONS_NAMES  # From config.py
         base_dictionary = {}  # Total dictionary{index -> action_name}
         for index, action in enumerate(action_names):
@@ -407,7 +350,9 @@ class TaxiEnv(environment_to_derive_from):
         if self.option_to_standby:
             available_action_list += ['turn_engine_on', 'turn_engine_off', 'standby']
 
-        if not self.max_fuel[0] == 0:
+        # TODO - when we return dictionary per taxi we can't longer assume that on np.inf fuel
+        #  means no limited fuel for all the taxis
+        if not self.max_fuel[0] == np.inf:
             available_action_list.append('refuel')
 
         action_index_dictionary = dict((value, key) for key, value in base_dictionary.items())  # {action -> index} all
@@ -423,6 +368,267 @@ class TaxiEnv(environment_to_derive_from):
         """
         return self.available_actions_indexes, self.index_action_dictionary
 
+    def _is_there_place_on_taxi(self, passengers_locations: np.array, taxi_index: int) -> bool:
+        """
+        Checks if there is room for another passenger on taxi number 'taxi_index'.
+        Args:
+            passengers_locations: list of all passengers locations
+            taxi_index: index of the desired taxi
+
+        Returns: Whether there is a place (True) or not (False)
+
+        """
+        # Remember that passengers "location" is: 1 - delivered, 2 - waits for a taxi, >2 - on a taxi with index
+        # location+2
+
+        return (len([location for location in passengers_locations if location == (taxi_index + 3)]) <
+                self.taxis_capacity[taxi_index])
+
+    def map_at_location(self, location: list) -> str:
+        """
+        Returns the map character on the specified coordinates of the grid.
+        Args:
+            location: location to check [row, col]
+
+        Returns: character on specific location on the map
+
+        """
+        domain_map = self.desc.copy().tolist()
+        row, col = location[0], location[1]
+        return domain_map[row + 1][2 * col + 1].decode(encoding='UTF-8')
+
+    def at_valid_fuel_station(self, taxi: int, taxis_locations: list) -> bool:
+        """
+        Checks if the taxi's location is a suitable fuel station or not.
+        Args:
+            taxi: the index of the desired taxi
+            taxis_locations: list of taxis coordinates [row, col]
+        Returns: whether the taxi is at a suitable fuel station (true) or not (false)
+
+        """
+        return (taxis_locations[taxi] in self.fuel_stations and
+                self.map_at_location(taxis_locations[taxi]) == self.fuel_type_list[taxi])
+
+    def _get_action_list(self, action_list) -> list:
+        """
+        Return a list in the correct format for the step function that should
+        always get a list even if it's a single action.
+        Args:
+            action_list:
+
+        Returns: list(action_list)
+
+        """
+        if type(action_list) == int:
+            return [action_list]
+        elif type(action_list) == np.int64:
+            return [action_list]
+
+        return action_list
+
+    def _engine_is_off_actions(self, action: str, taxi: int) -> int:
+        """
+        Returns the reward according to the requested action given that the engine's is currently off.
+        Also turns engine on if requested.
+        Args:
+            action: requested action
+            taxi: index of the taxi specified, relevant for turning engine on
+        Returns: correct reward
+
+        """
+        reward = self.partial_closest_path_reward('unrelated_action')
+        if action == 'standby':  # standby while engine is off
+            reward = self.partial_closest_path_reward('standby_engine_off')
+        elif action == 'turn_engine_on':  # turn engine on
+            reward = self.partial_closest_path_reward('turn_engine_on')
+            self.engine_status_list[taxi] = 1
+
+        return reward
+
+    def _take_movement(self, action: str, row: int, col: int) -> (bool, int, int):
+        """
+        Takes a movement with regard to a apecific location of a taxi,
+        Args:
+            action: direction to move
+            row: current row
+            col: current col
+
+        Returns: if moved (false if there is a wall), new row, new col
+
+        """
+        moved = False
+        new_row, new_col = row, col
+        max_row = self.num_rows - 1
+        max_col = self.num_columns - 1
+        if action == 'south':  # south
+            if row != max_row:
+                moved = True
+            new_row = min(row + 1, max_row)
+        elif action == 'north':  # north
+            if row != 0:
+                moved = True
+            new_row = max(row - 1, 0)
+        if action == 'east' and self.desc[1 + row, 2 * col + 2] == b":":  # east
+            if col != max_col:
+                moved = True
+            new_col = min(col + 1, max_col)
+        elif action == 'west' and self.desc[1 + row, 2 * col] == b":":  # west
+            if col != 0:
+                moved = True
+            new_col = max(col - 1, 0)
+
+        return moved, new_row, new_col
+
+    def _check_action_for_collision(self, taxi_index: int, taxis_locations: list, current_row: int, current_col: int,
+                                    moved: bool, current_action: int, current_reward: int) -> (int, bool, int, list):
+        """
+        Takes a desired location for a taxi and update it with regard to collision check.
+        Args:
+            taxi_index: index of the taxi
+            taxis_locations: locations of all other taxis.
+            current_row: of the taxi
+            current_col: of the taxi
+            moved: indicator variable
+            current_action: the current action requested
+            current_reward: the current reward (left unchanged if there is no collision)
+
+        Returns: new_reward, new_moved, new_action_index
+
+        """
+        reward = current_reward
+        row, col = current_row, current_col
+        moved = moved
+        action = current_action
+        taxi = taxi_index
+        # Check if the number of taxis on the destination location is greater than 0
+        if len([i for i in range(self.num_taxis) if taxis_locations[i] == [row, col]]) > 0:
+            if self.option_to_standby:
+                moved = False
+                action = self.action_index_dictionary['standby']
+            else:
+                self.collided[[i for i in range(len(taxis_locations)) if taxis_locations[i] == [row, col]]] = 1
+                self.collided[taxi] = 1
+                reward = self.partial_closest_path_reward('collision')
+                taxis_locations[taxi] = [row, col]
+
+        return reward, moved, action, taxis_locations
+
+    def _make_pickup(self, taxi: int, passengers_start_locations: list, passengers_status: list,
+                     taxi_location: list, reward: int) -> (list, int):
+        """
+        Make a pickup (successful or fail) for a given taxi.
+        Args:
+            taxi: index of the taxi
+            passengers_start_locations: current locations of the passengers
+            passengers_status: list of passengers statuses (1, 2, greater..)
+            taxi_location: location of the taxi
+            reward: current reward
+
+        Returns: updates passengers status list, updates reward
+
+        """
+        passengers_status = passengers_status
+        reward = reward
+        successful_pickup = False
+        for i, location in enumerate(passengers_status):
+            # Check if we can take this passenger
+            if location == 2 and taxi_location == passengers_start_locations[i] and \
+                    self._is_there_place_on_taxi(passengers_status, taxi):
+                passengers_status[i] = taxi + 3
+                successful_pickup = True
+                reward = self.partial_closest_path_reward('pickup')
+        if not successful_pickup:  # passenger not at location
+            reward = self.partial_closest_path_reward('bad_pickup')
+
+        return passengers_status, reward
+
+    def _make_dropoff(self, taxi: int, current_passengers_start_locations: list, current_passengers_status: list,
+                      destinations: list, taxi_location: list, reward: int) -> (list, list, int):
+        """
+        Make a dropoff (successful or fail) for a given taxi.
+        Args:
+            taxi: index of the taxi
+            current_passengers_start_locations: current locations of the passengers
+            current_passengers_status: list of passengers statuses (1, 2, greater..)
+            destinations: list of passengers destinations
+            taxi_location: location of the taxi
+            reward: current reward
+
+        Returns: updates passengers status list, updated passengers start location, updates reward
+
+        """
+        reward = reward
+        passengers_start_locations = current_passengers_start_locations.copy()
+        passengers_status = current_passengers_status.copy()
+        successful_dropoff = False
+        for i, location in enumerate(passengers_status):  # at destination
+            location = passengers_status[i]
+            # Check if we have the passenger and we are at his destination
+            if location == (taxi + 3) and taxi_location == destinations[i]:
+                passengers_status[i] = 1
+                reward = self.partial_closest_path_reward('final_dropoff', taxi)
+                passengers_start_locations[i] = taxi_location
+                successful_dropoff = True
+                break
+            elif location == (taxi + 3):  # drops off passenger not at destination
+                passengers_status[i] = 2
+                successful_dropoff = True
+                reward = self.partial_closest_path_reward('intermediate_dropoff', taxi)
+                passengers_start_locations[i] = taxi_location
+                break
+        if not successful_dropoff:  # not carrying a passenger
+            reward = self.partial_closest_path_reward('bad_dropoff')
+
+        return passengers_status, passengers_start_locations, reward
+
+    def _update_movement_wrt_fuel(self, taxi: int, taxis_locations: list, wanted_row: int, wanted_col: int,
+                                  reward: int, fuel: int) -> (int, int, list):
+        """
+        Given that a taxi would like to move - check the fuel accordingly and update reward and location.
+        Args:
+            taxi: index of the taxi
+            taxis_locations: list of current locations (prior to movement)
+            wanted_row: row after movement
+            wanted_col: col after movement
+            reward: current reward
+            fuel: current fuel
+
+        Returns: updated_reward, updated fuel, updared_taxis_locations
+
+        """
+        reward = reward
+        fuel = fuel
+        taxis_locations = taxis_locations
+        if fuel == 0:
+            reward = ('no_fuel')
+        else:
+            fuel = max(0, fuel - 1)
+
+            taxis_locations[taxi] = [wanted_row, wanted_col]
+
+        return reward, fuel, taxis_locations
+
+    def _refuel_taxi(self, current_fuel: int, current_reward: int, taxi: int, taxis_locations: list) -> (int, int):
+        """
+        Try to refuel a taxi, if successful - updates fuel tank, if not - updates the reward.
+        Args:
+            current_fuel: current fuel of the taxi
+            current_reward: current reward for the taxi.
+            taxi: taxi index
+            taxis_locations: list of current taxis locations
+
+        Returns: updated reward, updated fuel
+
+        """
+        fuel = current_fuel
+        reward = current_reward
+        if self.at_valid_fuel_station(taxi, taxis_locations) and fuel != self.max_fuel[taxi]:
+            fuel = self.max_fuel[taxi]
+        else:
+            reward = self.partial_closest_path_reward('bad_refuel')
+
+        return reward, fuel
+
     def step(self, action_dict: dict) -> (dict, dict, dict, dict):
         """
         Executing a list of actions (action for each taxi) at the domain current state.
@@ -435,15 +641,18 @@ class TaxiEnv(environment_to_derive_from):
         """
 
         rewards = {}
+        self.counter += 1
+        if self.counter >= 90:
+            self.window_size = 3
 
         # Main of the function, for each taxi-i act on action[i]
         for taxi_name, action_list in action_dict.items():
             # meta operations on the type of the action
-            action_list = basic_utils.get_action_list(action_list)
+            action_list = self._get_action_list(action_list)
 
             for action in action_list:
                 taxi = self.taxis_names.index(taxi_name)
-                reward = reward_utils.partial_closest_path_reward(self.state, 'step')  # Default reward
+                reward = self.partial_closest_path_reward('step')  # Default reward
                 moved = False  # Indicator variable for later use
 
                 # taxi locations: [i, j]
@@ -453,20 +662,18 @@ class TaxiEnv(environment_to_derive_from):
                 taxis_locations, fuels, passengers_start_locations, destinations, passengers_status = self.state
 
                 if all(list(self.dones.values())):
+                    rewards[taxi_name] = reward
                     continue
 
                 # If taxi is collided, it can't perform a step
                 if self.collided[taxi] == 1:
-                    rewards[taxi_name] = reward_utils.partial_closest_path_reward(self.state, 'collided')
+                    rewards[taxi_name] = self.partial_closest_path_reward('collided')
                     self.dones[taxi_name] = True
                     continue
 
                 # If the taxi is out of fuel, it can't perform a step
-                if fuels[taxi] == 0 and not basic_utils.at_valid_fuel_station(taxi, taxis_locations, self.fuel_stations,
-                                                                              self.desc.copy().tolist(),
-                                                                              self.fuel_type_list) and not \
-                        self.is_infinite_fuel:
-                    rewards[taxi_name] = reward_utils.partial_closest_path_reward(self.state, 'bad_fuel')
+                if fuels[taxi] == 0 and not self.at_valid_fuel_station(taxi, taxis_locations):
+                    rewards[taxi_name] = self.partial_closest_path_reward('bad_fuel')
                     self.dones[taxi_name] = True
                     continue
 
@@ -479,70 +686,55 @@ class TaxiEnv(environment_to_derive_from):
 
                 if not is_taxi_engine_on:  # Engine is off
                     # update reward according to standby/ turn-on/ unrelated + turn engine on if requsted
-                    reward, self.engine_status_list = \
-                        actions_utils.engine_is_off_actions(self.state, index_action_dictionary[action], taxi,
-                                                            reward_utils.partial_closest_path_reward,
-                                                            self.engine_status_list)
+                    reward = self._engine_is_off_actions(index_action_dictionary[action], taxi)
 
 
                 else:  # Engine is on
                     # Binding
                     if index_action_dictionary[action] == 'bind':
                         self.bounded = False
-                        reward = reward_utils.partial_closest_path_reward(self.state, 'bind')
+                        reward = self.partial_closest_path_reward('bind')
 
                     # Movement
                     if index_action_dictionary[action] in ['south', 'north', 'east', 'west']:
-                        moved, row, col = actions_utils.take_movement(index_action_dictionary[action], row, col,
-                                                                      self.num_rows, self.num_columns, self.desc)
+                        moved, row, col = self._take_movement(index_action_dictionary[action], row, col)
 
                     # Check for collisions
                     if self.collision_sensitive_domain and moved:
                         if self.collided[taxi] == 0:
-                            reward, moved, action, taxis_locations, self.collided = \
-                                actions_utils.check_action_for_collision(self.state, taxi, taxis_locations, row, col,
-                                                                         moved, action, reward, self.num_taxis,
-                                                                         self.option_to_standby,
-                                                                         self.action_index_dictionary, self.collided,
-                                                                         reward_utils.partial_closest_path_reward)
+                            reward, moved, action, taxis_locations = self._check_action_for_collision(taxi,
+                                                                                                      taxis_locations,
+                                                                                                      row, col, moved,
+                                                                                                      action, reward)
 
                     # Pickup
                     elif index_action_dictionary[action] == 'pickup':
-                        passengers_status, reward = actions_utils.make_pickup(self.state, taxi,
-                                                                              passengers_start_locations,
-                                                                              passengers_status, taxi_location, reward,
-                                                                              self.taxis_capacity,
-                                                                              reward_utils.partial_closest_path_reward)
+                        passengers_status, reward = self._make_pickup(taxi, passengers_start_locations,
+                                                                      passengers_status, taxi_location, reward)
 
                     # Dropoff
-                    elif 'dropoff' in index_action_dictionary[action]:
-                        passenger_index = int(index_action_dictionary[action][7:])
-                        passengers_status, passengers_start_locations, reward = actions_utils.make_dropoff(self.state,
-                                                                                                           taxi,
-                                                                                                           passengers_start_locations,
-                                                                                                           passengers_status,
-                                                                                                           destinations,
-                                                                                                           taxi_location,
-                                                                                                           reward,
-                                                                                                           passenger_index,
-                                                                                                           reward_utils.partial_closest_path_reward)
+                    elif index_action_dictionary[action] == 'dropoff':
+                        passengers_status, passengers_start_locations, reward = self._make_dropoff(taxi,
+                                                                                                   passengers_start_locations,
+                                                                                                   passengers_status,
+                                                                                                   destinations,
+                                                                                                   taxi_location,
+                                                                                                   reward)
 
                     # Turning engine off
                     elif index_action_dictionary[action] == 'turn_engine_off':
-                        reward = reward_utils.partial_closest_path_reward(self.state, 'turn_engine_off')
+                        reward = self.partial_closest_path_reward('turn_engine_off')
                         self.engine_status_list[taxi] = 0
 
                     # Standby with engine on
                     elif index_action_dictionary[action] == 'standby':
-                        reward = reward_utils.partial_closest_path_reward(self.state, 'standby_engine_on')
+                        reward = self.partial_closest_path_reward('standby_engine_on')
 
                 # Here we have finished checking for action for taxi-i
                 # Fuel consumption
                 if moved:
-                    reward, fuels[taxi], taxis_locations = actions_utils.update_movement_wrt_fuel(self.state, taxi,
-                                                                                                  taxis_locations,
-                                                                                                  row, col, reward, fuel, self.is_infinite_fuel,
-                                                                                                  reward_utils.partial_closest_path_reward)
+                    reward, fuels[taxi], taxis_locations = self._update_movement_wrt_fuel(taxi, taxis_locations,
+                                                                                          row, col, reward, fuel)
 
                 if (not moved) and action in [self.action_index_dictionary[direction] for
                                               direction in ['north', 'south', 'west', 'east']]:
@@ -550,11 +742,7 @@ class TaxiEnv(environment_to_derive_from):
 
                 # taxi refuel
                 if index_action_dictionary[action] == 'refuel':
-                    reward, fuels[taxi] = actions_utils.refuel_taxi(self.state, fuel, reward, taxi, taxis_locations,
-                                                                    self.fuel_stations, self.desc.copy().tolist(),
-                                                                    self.fuel_type_list, self.is_infinite_fuel,
-                                                                    self.max_fuel,
-                                                                    reward_utils.partial_closest_path_reward)
+                    reward, fuels[taxi] = self._refuel_taxi(fuel, reward, taxi, taxis_locations)
 
                 # check if all the passengers are at their destinations
                 done = all(loc == 1 for loc in passengers_status)
@@ -565,7 +753,7 @@ class TaxiEnv(environment_to_derive_from):
                 self.dones[taxi_name] = self.dones[taxi_name] or done
 
                 # check if all taxis are out of fuel
-                done = fuels[taxi] == 0 and not self.is_infinite_fuel
+                done = fuels[taxi] == 0
                 self.dones[taxi_name] = self.dones[taxi_name] or done
 
                 rewards[taxi_name] = reward
@@ -574,9 +762,6 @@ class TaxiEnv(environment_to_derive_from):
 
         self.dones['__all__'] = True
         self.dones['__all__'] = all(list(self.dones.values()))
-
-        # rendering_functions.render(self.desc.copy().tolist(), self.state, self.num_taxis, self.collided,
-        #                            self.last_action, self.action_index_dictionary, self.dones)
 
         if self.bounded:
             total_reward = 0
@@ -587,16 +772,204 @@ class TaxiEnv(environment_to_derive_from):
                 rewards[taxi_id] = total_reward
 
         obs = {}
-        for i, taxi_id in enumerate(action_dict.keys()):
-            # obs[taxi_id] = self.get_observation(self.state, taxi_id)
-            obs[taxi_id] = (observation_utils.get_image_obs_by_agent_id(i, self.state, self.num_taxis,
-                                                                        self.collided, self.view_len,
-                                                                        self.desc.copy()),
-                            observation_utils.get_status_vector_observation(self.state, taxi_id, self.taxis_names,
-                                                                            self.num_taxis))
-        # im = rendering_functions.map2rgb(self.state, self.desc.astype(str))
-        # plt.imshow(im)
-        # rendering_functions.plot_window_of_observation(self.view_len, self.state[0])
-        # plt.show()
-        # return obs, {taxi_id: rewards[taxi_id] for taxi_id in action_dict.keys()}, self.dones, {}
-        return obs, rewards, self.dones, {}
+        for taxi_id in action_dict.keys():
+            obs[taxi_id] = self.get_observation(self.state, taxi_id)
+
+        return obs, {taxi_id: rewards[taxi_id] for taxi_id in action_dict.keys()}, self.dones, {}
+
+    def render(self, mode: str = 'human') -> str:
+        """
+        Renders the domain map at the current state
+        Args:
+            mode: Demand mode (file or human watching).
+
+        Returns: Value string of writing the output
+
+        """
+        outfile = StringIO() if mode == 'ansi' else sys.stdout
+
+        # Copy map to work on
+        out = self.desc.copy().tolist()
+        out = [[c.decode('utf-8') for c in line] for line in out]
+
+        taxis, fuels, passengers_start_coordinates, destinations, passengers_locations = self.state
+
+        colors = ['yellow', 'red', 'white', 'green', 'cyan', 'crimson', 'gray', 'magenta'] * 5
+        colored = [False] * self.num_taxis
+
+        def ul(x):
+            """returns underline instead of spaces when called"""
+            return "_" if x == " " else x
+
+        for i, location in enumerate(passengers_locations):
+            if location > 2:  # Passenger is on a taxi
+                taxi_row, taxi_col = taxis[location - 3]
+
+                # Coloring taxi's coordinate on the map
+                out[1 + taxi_row][2 * taxi_col + 1] = utils.colorize(
+                    out[1 + taxi_row][2 * taxi_col + 1], colors[location - 3], highlight=True, bold=True)
+                colored[location - 3] = True
+            else:  # Passenger isn't in a taxi
+                # Coloring passenger's coordinates on the map
+                pi, pj = passengers_start_coordinates[i]
+                out[1 + pi][2 * pj + 1] = utils.colorize(out[1 + pi][2 * pj + 1], 'blue', bold=True)
+
+        for i, taxi in enumerate(taxis):
+            if self.collided[i] == 0:  # Taxi isn't collided
+                taxi_row, taxi_col = taxi
+                out[1 + taxi_row][2 * taxi_col + 1] = utils.colorize(
+                    ul(out[1 + taxi_row][2 * taxi_col + 1]), colors[i], highlight=True)
+            else:  # Collided!
+                taxi_row, taxi_col = taxi
+                out[1 + taxi_row][2 * taxi_col + 1] = utils.colorize(
+                    ul(out[1 + taxi_row][2 * taxi_col + 1]), 'gray', highlight=True)
+
+        for dest in destinations:
+            di, dj = dest
+            out[1 + di][2 * dj + 1] = utils.colorize(out[1 + di][2 * dj + 1], 'magenta')
+        outfile.write("\n".join(["".join(row) for row in out]) + "\n")
+
+        if self.last_action is not None:
+            moves = ALL_ACTIONS_NAMES
+            output = [moves[i] for i in np.array(list(self.last_action.values())).reshape(-1)]
+            outfile.write("  ({})\n".format(' ,'.join(output)))
+        for i, taxi in enumerate(taxis):
+            outfile.write("Taxi{}-{}: Fuel: {}, Location: ({},{}), Collided: {}\n".format(i + 1, colors[i].upper(),
+                                                                                          fuels[i], taxi[0], taxi[1],
+                                                                                          self.collided[i] == 1))
+        for i, location in enumerate(passengers_locations):
+            start = tuple(passengers_start_coordinates[i])
+            end = tuple(destinations[i])
+            if location == 1:
+                outfile.write("Passenger{}: Location: Arrived!, Destination: {}\n".format(i + 1, end))
+            if location == 2:
+                outfile.write("Passenger{}: Location: {}, Destination: {}\n".format(i + 1, start, end))
+            else:
+                outfile.write("Passenger{}: Location: Taxi{}, Destination: {}\n".format(i + 1, location - 2, end))
+        outfile.write("Done: {}, {}\n".format(all(self.dones.values()), self.dones))
+        outfile.write("Passengers Status's: {}\n".format(self.state[-1]))
+
+        # No need to return anything for human
+        if mode != 'human':
+            with closing(outfile):
+                return outfile.getvalue()
+
+    @staticmethod
+    def partial_observations(state: list) -> list:
+        """
+        Get partial observation of state.
+        Args:
+            state: state of the domain (taxis, fuels, passengers_start_coordinates, destinations, passengers_locations)
+
+        Returns: list of observations s.t each taxi sees only itself
+
+        """
+
+        def flatten(x):
+            return [item for sub in x for item in sub]
+
+        observations = []
+        taxis, fuels, passengers_start_locations, passengers_destinations, passengers_locations = state
+        pass_info = flatten(passengers_start_locations) + flatten(passengers_destinations) + passengers_locations
+
+        for i in range(len(taxis)):
+            obs = taxis[i] + [fuels[i]] + pass_info
+            obs = np.reshape(obs, [1, len(obs)])
+            observations.append(obs)
+        return observations
+
+    def get_l1_distance(self, location1, location2):
+        """
+        Return the minimal travel length between 2 locations on the grid world.
+        Args:
+            location1: [i1, j1]
+            location2: [i2, j2]
+
+        Returns: np.abs(i1 - i2) + np.abs(j1 - j2)
+
+        """
+        return np.abs(location1[0] - location2[0]) + np.abs(location1[1] - location2[1])
+
+    def get_observation(self, state: list, agent_name: str) -> np.array:
+        """
+        Takes only the observation of the specified agent.
+        Args:
+            state: state of the domain (taxis, fuels, passengers_start_coordinates, destinations, passengers_locations)
+            agent_name: observer name
+            window_size: the size that the agent can see in the map (around it) in terms of other txis
+
+        Returns: observation of the specified agent (state wise)
+
+        """
+
+        def flatten(x):
+            return [item for sub in list(x) for item in list(sub)]
+
+        agent_index = self.taxis_names.index(agent_name)
+
+        taxis, fuels, passengers_start_locations, passengers_destinations, passengers_locations = state.copy()
+        passengers_information = flatten(passengers_start_locations) + flatten(
+            passengers_destinations) + passengers_locations
+
+        closest_taxis_indices = []
+        for i in range(self.num_taxis):
+            if self.get_l1_distance(taxis[agent_index], taxis[i]) <= self.window_size and i != agent_index:
+                closest_taxis_indices.append(i)
+
+        observations = taxis[agent_index].copy()
+        for i in closest_taxis_indices:
+            observations += taxis[i]
+        observations += [0, 0] * (self.num_taxis - 1 - len(closest_taxis_indices)) + [fuels[agent_index]] + \
+                        [0] * (self.num_taxis - 1) + passengers_information
+        observations = np.reshape(observations, (1, len(observations)))
+
+        return observations
+
+    def passenger_destination_l1_distance(self, passenger_index, current_row: int, current_col: int) -> int:
+        """
+        Returns the manhattan distance between passenger current defined "start location" and it's destination.
+        Args:
+            passenger_index: index of the passenger.
+            current_row: current row to calculate distance from destination
+            current_col: current col to calculate distance from destination
+
+        Returns: manhattan distance
+
+        """
+        current_state = self.state
+        destination_row, destination_col = current_state[3][passenger_index]
+        return int(np.abs(current_col - destination_col) + np.abs(current_row - destination_row))
+
+    def partial_closest_path_reward(self, basic_reward_str: str, taxi_index: int = None) -> int:
+        """
+        Computes the reward for a taxi and it's defined by:
+        dropoff[s] - gets the reward equal to the closest path multiply by 15, if the drive got a passenger further
+        away - negative.
+        other actions - basic reward from config table
+        Args:
+            basic_reward_str: the reward we would like to give
+            taxi_index: index of the specific taxi
+
+        Returns: updated reward
+
+        """
+        if basic_reward_str not in ['intermediate_dropoff', 'final_dropoff'] or taxi_index is None:
+            return TAXI_ENVIRONMENT_REWARDS[basic_reward_str]
+
+        # [taxis_locations, fuels, passengers_start_locations, destinations, passengers_status]
+        current_state = self.state
+        passengers_start_locations = current_state[2]
+
+        taxis_locations = current_state[0]
+
+        passengers_status = current_state[-1]
+        passenger_index = passengers_status.index(taxi_index + 3)
+        passenger_start_row, passenger_start_col = passengers_start_locations[passenger_index]
+        taxi_current_row, taxi_current_col = taxis_locations[taxi_index]
+
+        return 15 * (self.passenger_destination_l1_distance(passenger_index, passenger_start_row, passenger_start_col) -
+                     self.passenger_destination_l1_distance(passenger_index, taxi_current_row, taxi_current_col))
+
+
+env = TaxiEnv()
+env.render()
